pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "ai-assistant"
version = "0.1.0"
description = "AI Assistant project"
authors = ["ebrahim saad <ebrahimsaad2017@hotmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://example.com"
repository = "https://github.com/your-repo/ai-assistant"
keywords = ["AI", "assistant"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "0.115.12"
uvicorn = "0.34.2"
pydantic-settings = "2.9.1"
langchain = "0.3.24"
langchain-community = "0.3.23"
langchain-together = "0.3.0"
langchain-tavily = "0.1.5"
langchainhub = "0.1.21"
langgraph = "0.3.34"
json-repair = "0.43.0"
tavily-python = "0.7.0"
tiktoken = "0.9.0"
python-multipart = "0.0.20"
aiofiles = "24.1.0"
PyMuPDF = "1.25.5"
motor = "3.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
black = "*"
isort = "*"
flake8 = "*"
